refactor(cli): update test imports and remove migrated scripts
Port builder functions missing from the new CLI modules (build_logs_command,
build_whisplay_command, build_rtc_command, build_parser, run_screenshot) into
yoyopy/cli/remote/ops.py. Update test_pi_remote.py and test_whisplay_tune.py
to import from yoyopy.cli.remote.* and yoyopy.cli.pi.tune instead of the old
scripts/ module paths. Delete all migrated scripts, keeping only
scripts/generate_test_sounds.py.
