--- scripts/pi_remote.py
@@ -1,1581 +1,0 @@
-#!/usr/bin/env python3
-"""Developer helper for common YoyoPod Raspberry Pi workflows over SSH."""
-
-from __future__ import annotations
-
-import argparse
-import fnmatch
-import json
-import os
-import shlex
-import shutil
-import subprocess
-import sys
-import tarfile
-import tempfile
-from dataclasses import dataclass
-from pathlib import Path
-from typing import Sequence
-
-import yaml
-
-
-REPO_ROOT = Path(__file__).resolve().parent.parent
-DEPLOY_CONFIG_PATH = REPO_ROOT / "deploy" / "pi-deploy.yaml"
-LOCAL_DEPLOY_CONFIG_PATH = REPO_ROOT / "deploy" / "pi-deploy.local.yaml"
-
-
-@dataclass
-class RemoteConfig:
-    """Connection details for the Raspberry Pi host."""
-
-    host: str
-    user: str
-    project_dir: str
-    branch: str
-
-    @property
-    def ssh_target(self) -> str:
-        """Return the SSH target in user@host form when a user is configured."""
-        if self.user:
-            return f"{self.user}@{self.host}"
-        return self.host
-
-
-@dataclass(frozen=True)
-class PiDeployConfig:
-    """Stable runtime paths used by the Pi deploy/debugging workflow."""
-
-    host: str = ""
-    user: str = ""
-    project_dir: str = "~/yoyo-py"
-    branch: str = "main"
-    venv: str = ".venv"
-    start_cmd: str = "python yoyopod.py"
-    kill_processes: tuple[str, ...] = ("python", "linphonec")
-    log_file: str = "logs/yoyopod.log"
-    error_log_file: str = "logs/yoyopod_errors.log"
-    pid_file: str = "/tmp/yoyopod.pid"
-    startup_marker: str = "YoyoPod starting"
-    screenshot_path: str = "/tmp/yoyopod_screenshot.png"
-    rsync_exclude: tuple[str, ...] = (
-        ".git/",
-        ".cache/",
-        "__pycache__/",
-        "*.pyc",
-        ".venv/",
-        "build/",
-        "logs/",
-        "node_modules/",
-        "*.egg-info/",
-    )
-
-
-def load_yaml_mapping(path: Path) -> dict[str, object]:
-    """Load one YAML mapping from disk."""
-
-    with open(path, "r", encoding="utf-8") as handle:
-        data = yaml.safe_load(handle) or {}
-    if not isinstance(data, dict):
-        raise SystemExit(f"Expected a YAML mapping in {path}")
-    return data
-
-
-def merge_pi_deploy_layers(*layers: dict[str, object]) -> dict[str, object]:
-    """Merge deploy config layers from lowest to highest precedence."""
-
-    merged: dict[str, object] = {}
-    for layer in layers:
-        for key, value in layer.items():
-            if value is not None:
-                merged[key] = value
-    return merged
-
-
-def parse_pi_deploy_config(data: dict[str, object]) -> PiDeployConfig:
-    """Normalize raw YAML data into a deploy config object."""
-
-    return PiDeployConfig(
-        host=str(data.get("host", "")).strip(),
-        user=str(data.get("user", "")).strip(),
-        project_dir=str(
-            data.get("project_dir", data.get("remote_dir", "~/yoyo-py"))
-        ).strip()
-        or "~/yoyo-py",
-        branch=str(data.get("branch", "main")).strip() or "main",
-        venv=str(data.get("venv", ".venv")).strip() or ".venv",
-        start_cmd=str(data.get("start_cmd", "python yoyopod.py")).strip()
-        or "python yoyopod.py",
-        kill_processes=tuple(
-            str(process).strip()
-            for process in data.get("kill_processes", ("python", "linphonec"))
-            if str(process).strip()
-        )
-        or ("python", "linphonec"),
-        log_file=str(data["log_file"]).strip(),
-        error_log_file=str(data["error_log_file"]).strip(),
-        pid_file=str(data["pid_file"]).strip(),
-        startup_marker=str(data["startup_marker"]).strip(),
-        screenshot_path=str(
-            data.get("screenshot_path", "/tmp/yoyopod_screenshot.png")
-        ).strip()
-        or "/tmp/yoyopod_screenshot.png",
-        rsync_exclude=tuple(
-            str(pattern)
-            for pattern in data.get(
-                "rsync_exclude",
-                (
-                    ".git/",
-                    ".cache/",
-                    "__pycache__/",
-                    "*.pyc",
-                    ".venv/",
-                    "build/",
-                    "logs/",
-                    "node_modules/",
-                    "*.egg-info/",
-                ),
-            )
-        ),
-    )
-
-
-def load_pi_deploy_config(
-    *,
-    config_path: Path | None = None,
-    local_override_path: Path | None = None,
-) -> PiDeployConfig:
-    """Load the shared deploy config with an optional local override layer."""
-
-    base_path = config_path or DEPLOY_CONFIG_PATH
-    local_path = local_override_path or LOCAL_DEPLOY_CONFIG_PATH
-
-    merged_data = load_yaml_mapping(base_path)
-    if local_path.exists():
-        merged_data = merge_pi_deploy_layers(
-            merged_data,
-            load_yaml_mapping(local_path),
-        )
-
-    return parse_pi_deploy_config(merged_data)
-
-
-def pi_deploy_config_to_dict(config: PiDeployConfig) -> dict[str, object]:
-    """Convert one deploy config object back into a YAML-friendly mapping."""
-
-    return {
-        "host": config.host,
-        "user": config.user,
-        "project_dir": config.project_dir,
-        "branch": config.branch,
-        "venv": config.venv,
-        "start_cmd": config.start_cmd,
-        "kill_processes": list(config.kill_processes),
-        "log_file": config.log_file,
-        "error_log_file": config.error_log_file,
-        "pid_file": config.pid_file,
-        "startup_marker": config.startup_marker,
-        "screenshot_path": config.screenshot_path,
-        "rsync_exclude": list(config.rsync_exclude),
-    }
-
-
-def build_local_override_template(base_config: PiDeployConfig) -> str:
-    """Create the starter template for the gitignored local override file."""
-
-    host = base_config.host or "rpi-zero"
-    user = base_config.user or "pi"
-    body = yaml.safe_dump(
-        {
-            "host": host,
-            "user": user,
-            "project_dir": base_config.project_dir,
-            "branch": base_config.branch,
-        },
-        sort_keys=False,
-    ).rstrip()
-    return (
-        "# Local Raspberry Pi overrides for this workstation.\n"
-        "# This file is gitignored. Only machine-specific defaults belong here.\n"
-        "# Precedence is: deploy/pi-deploy.yaml -> deploy/pi-deploy.local.yaml -> env -> CLI.\n"
-        f"{body}\n"
-    )
-
-
-def ensure_local_pi_deploy_config(
-    base_config: PiDeployConfig,
-    *,
-    local_override_path: Path | None = None,
-) -> tuple[Path, bool]:
-    """Create the gitignored local override file when it does not exist yet."""
-
-    local_path = local_override_path or LOCAL_DEPLOY_CONFIG_PATH
-    if local_path.exists():
-        return local_path, False
-
-    local_path.parent.mkdir(parents=True, exist_ok=True)
-    local_path.write_text(
-        build_local_override_template(base_config),
-        encoding="utf-8",
-    )
-    return local_path, True
-
-
-def build_config_editor_command(
-    config_path: Path,
-    *,
-    editor: str | None = None,
-) -> list[str]:
-    """Resolve the best local editor command for the override file."""
-
-    configured_editor = editor or os.getenv("VISUAL") or os.getenv("EDITOR")
-    if configured_editor:
-        return [*shlex.split(configured_editor), str(config_path)]
-
-    if sys.platform.startswith("win"):
-        return ["notepad", str(config_path)]
-
-    if sys.platform == "darwin":
-        return ["open", "-W", "-t", str(config_path)]
-
-    for candidate in ("sensible-editor", "nano", "vi", "xdg-open"):
-        if shutil.which(candidate):
-            return [candidate, str(config_path)]
-
-    return ["xdg-open", str(config_path)]
-
-
-def build_parser(deploy_config: PiDeployConfig) -> argparse.ArgumentParser:
-    """Create the command-line parser."""
-    parser = argparse.ArgumentParser(
-        description=(
-            "Run common YoyoPod Raspberry Pi development tasks over SSH. "
-            "Defaults can be provided with YOYOPOD_PI_HOST, "
-            "YOYOPOD_PI_USER, YOYOPOD_PI_PROJECT_DIR, and YOYOPOD_PI_BRANCH, "
-            "or through deploy/pi-deploy.yaml plus the optional "
-            "deploy/pi-deploy.local.yaml override."
-        )
-    )
-    parser.add_argument(
-        "--host",
-        default=os.getenv("YOYOPOD_PI_HOST", deploy_config.host),
-        help="SSH host or alias for the Raspberry Pi",
-    )
-    parser.add_argument(
-        "--user",
-        default=os.getenv("YOYOPOD_PI_USER", deploy_config.user),
-        help="SSH user for the Raspberry Pi (optional)",
-    )
-    parser.add_argument(
-        "--project-dir",
-        default=os.getenv("YOYOPOD_PI_PROJECT_DIR", deploy_config.project_dir),
-        help="Project directory on the Raspberry Pi (default: ~/yoyo-py)",
-    )
-    parser.add_argument(
-        "--branch",
-        default=os.getenv("YOYOPOD_PI_BRANCH", deploy_config.branch),
-        help="Git branch to sync on the Raspberry Pi (default: main)",
-    )
-
-    subparsers = parser.add_subparsers(dest="command", required=True)
-
-    config_parser = subparsers.add_parser(
-        "config",
-        help="Show or edit the merged Raspberry Pi deploy config",
-    )
-    config_parser.add_argument(
-        "config_action",
-        nargs="?",
-        default="show",
-        choices=["show", "paths", "init-local", "edit"],
-        help="Config action to run locally (default: show)",
-    )
-    config_parser.add_argument(
-        "--editor",
-        help="Override the editor command for `config edit`",
-    )
-
-    subparsers.add_parser(
-        "status",
-        help="Show remote repo, music backend, and process status",
-    )
-
-    sync_parser = subparsers.add_parser(
-        "sync",
-        help="Fetch, checkout, pull, and optionally run uv sync on the Raspberry Pi",
-    )
-    sync_parser.add_argument(
-        "--skip-uv-sync",
-        action="store_true",
-        help="Skip `uv sync --extra dev` after pulling",
-    )
-
-    rsync_parser = subparsers.add_parser(
-        "rsync",
-        help="Rsync the local working tree to the Raspberry Pi and restart the app",
-    )
-    rsync_parser.add_argument(
-        "--skip-restart",
-        action="store_true",
-        help="Only rsync files without restarting the app afterward",
-    )
-
-    restart_parser = subparsers.add_parser(
-        "restart",
-        help="Kill and relaunch the production app on the Raspberry Pi",
-    )
-
-    screenshot_parser = subparsers.add_parser(
-        "screenshot",
-        help="Capture a screenshot from the running app on the Raspberry Pi",
-    )
-    screenshot_parser.add_argument(
-        "--readback",
-        action="store_true",
-        help="Use LVGL readback/default capture (SIGUSR1) instead of the legacy shadow-first path (SIGUSR2)",
-    )
-    screenshot_parser.add_argument(
-        "--output",
-        default="pi_screenshot.png",
-        help="Local path for the downloaded screenshot (default: ./pi_screenshot.png)",
-    )
-
-    smoke_parser = subparsers.add_parser(
-        "smoke",
-        help="Run the Raspberry Pi smoke validator remotely",
-    )
-    smoke_parser.add_argument(
-        "--with-power",
-        action="store_true",
-        help="Include PiSugar power checks",
-    )
-    smoke_parser.add_argument(
-        "--with-rtc",
-        action="store_true",
-        help="Include PiSugar RTC checks",
-    )
-    smoke_parser.add_argument(
-        "--with-music",
-        action="store_true",
-        help="Include music-backend startup checks",
-    )
-    smoke_parser.add_argument(
-        "--with-voip",
-        action="store_true",
-        help="Include SIP registration checks",
-    )
-    smoke_parser.add_argument(
-        "--with-lvgl-soak",
-        action="store_true",
-        help="Include a short LVGL transition and sleep/wake soak",
-    )
-    smoke_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose smoke-script logging",
-    )
-    smoke_parser.add_argument(
-        "--music-timeout",
-        type=int,
-        default=5,
-        help="Music-backend startup timeout in seconds (default: 5)",
-    )
-    smoke_parser.add_argument(
-        "--voip-timeout",
-        type=float,
-        default=10.0,
-        help="VoIP registration timeout in seconds (default: 10)",
-    )
-
-    whisplay_parser = subparsers.add_parser(
-        "whisplay",
-        help="Run the Whisplay gesture-tuning helper remotely",
-    )
-    whisplay_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose tuner logging",
-    )
-    whisplay_parser.add_argument(
-        "--duration-seconds",
-        type=float,
-        default=30.0,
-        help="How long to monitor gestures before exiting (default: 30)",
-    )
-    whisplay_parser.add_argument(
-        "--debounce-ms",
-        type=int,
-        help="Temporary Whisplay debounce override in milliseconds",
-    )
-    whisplay_parser.add_argument(
-        "--double-tap-ms",
-        type=int,
-        help="Temporary Whisplay double-tap override in milliseconds",
-    )
-    whisplay_parser.add_argument(
-        "--long-hold-ms",
-        type=int,
-        help="Temporary Whisplay long-hold override in milliseconds",
-    )
-    whisplay_parser.add_argument(
-        "--no-display",
-        action="store_true",
-        help="Log events only without drawing tuner hints on the display",
-    )
-
-    lvgl_soak_parser = subparsers.add_parser(
-        "lvgl-soak",
-        help="Run the LVGL Whisplay soak helper remotely",
-    )
-    lvgl_soak_parser.add_argument(
-        "--cycles",
-        type=int,
-        default=2,
-        help="How many full transition cycles to run (default: 2)",
-    )
-    lvgl_soak_parser.add_argument(
-        "--hold-seconds",
-        type=float,
-        default=0.2,
-        help="How long to keep each screen active (default: 0.2)",
-    )
-    lvgl_soak_parser.add_argument(
-        "--skip-sleep",
-        action="store_true",
-        help="Skip the sleep/wake exercise",
-    )
-    lvgl_soak_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose soak logging",
-    )
-
-    rtc_parser = subparsers.add_parser(
-        "rtc",
-        help="Inspect or control PiSugar RTC state remotely",
-    )
-    rtc_parser.add_argument(
-        "rtc_action",
-        nargs="?",
-        default="status",
-        choices=["status", "sync-to-rtc", "sync-from-rtc", "set-alarm", "disable-alarm"],
-        help="RTC action to run remotely (default: status)",
-    )
-    rtc_parser.add_argument(
-        "--time",
-        help="ISO8601 timestamp for set-alarm, e.g. 2026-04-06T07:30:00+02:00",
-    )
-    rtc_parser.add_argument(
-        "--repeat-mask",
-        type=int,
-        default=127,
-        help="Weekday repeat bitmask for set-alarm (default: 127)",
-    )
-    rtc_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose RTC helper logging",
-    )
-
-    power_parser = subparsers.add_parser(
-        "power",
-        help="Inspect PiSugar power telemetry remotely",
-    )
-    power_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose power helper logging",
-    )
-
-    logs_parser = subparsers.add_parser(
-        "logs",
-        help="Tail the file-based YoyoPod logs on the Raspberry Pi",
-    )
-    logs_parser.add_argument(
-        "--errors",
-        action="store_true",
-        help="Read the errors-only log file instead of the main log",
-    )
-    logs_parser.add_argument(
-        "--follow",
-        action="store_true",
-        help="Follow the log output until interrupted",
-    )
-    logs_parser.add_argument(
-        "--filter",
-        help="Case-insensitive grep filter for subsystem, module, level, or text",
-    )
-    logs_parser.add_argument(
-        "--lines",
-        type=int,
-        default=100,
-        help="How many lines to show before following (default: 100)",
-    )
-
-    service_parser = subparsers.add_parser(
-        "service",
-        help="Install or inspect the production YoyoPod systemd service",
-    )
-    service_parser.add_argument(
-        "service_action",
-        nargs="?",
-        default="status",
-        choices=["status", "install", "start", "stop", "restart", "logs"],
-        help="Service action to run remotely (default: status)",
-    )
-    service_parser.add_argument(
-        "--lines",
-        type=int,
-        default=100,
-        help="How many journal lines to show for `service logs` (default: 100)",
-    )
-
-    preflight_parser = subparsers.add_parser(
-        "preflight",
-        help="Run local checks, sync the Pi, and execute the Pi smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--skip-local",
-        action="store_true",
-        help="Skip local compile/test verification before remote work",
-    )
-    preflight_parser.add_argument(
-        "--skip-sync",
-        action="store_true",
-        help="Skip the remote git pull and dependency sync step",
-    )
-    preflight_parser.add_argument(
-        "--skip-uv-sync",
-        action="store_true",
-        help="Skip `uv sync --extra dev` during the remote sync step",
-    )
-    preflight_parser.add_argument(
-        "--with-power",
-        action="store_true",
-        help="Include PiSugar power checks in the remote smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--with-rtc",
-        action="store_true",
-        help="Include PiSugar RTC checks in the remote smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--with-music",
-        action="store_true",
-        help="Include music-backend startup checks in the remote smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--with-voip",
-        action="store_true",
-        help="Include SIP registration checks in the remote smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--with-lvgl-soak",
-        action="store_true",
-        help="Include the LVGL soak helper in the remote smoke pass",
-    )
-    preflight_parser.add_argument(
-        "--verbose",
-        action="store_true",
-        help="Enable verbose smoke-script logging",
-    )
-    preflight_parser.add_argument(
-        "--music-timeout",
-        type=int,
-        default=5,
-        help="Music-backend startup timeout in seconds (default: 5)",
-    )
-    preflight_parser.add_argument(
-        "--voip-timeout",
-        type=float,
-        default=10.0,
-        help="VoIP registration timeout in seconds (default: 10)",
-    )
-
-    run_parser = subparsers.add_parser(
-        "run",
-        help="Start the production app on the Raspberry Pi",
-    )
-    run_parser.add_argument(
-        "--simulate",
-        action="store_true",
-        help="Run the app in simulation mode instead of hardware mode",
-    )
-    run_parser.add_argument(
-        "--app-arg",
-        action="append",
-        default=[],
-        help="Extra argument to pass through to yoyopod.py (repeatable)",
-    )
-
-    return parser
-
-
-def validate_config(config: RemoteConfig) -> None:
-    """Ensure required connection details are present."""
-    if not config.host:
-        raise SystemExit(
-            "Missing Raspberry Pi host. Set it with "
-            "`uv run python scripts/pi_remote.py config edit`, "
-            "pass --host, or set YOYOPOD_PI_HOST."
-        )
-
-
-def quote_remote_project_dir(project_dir: str) -> str:
-    """Quote the remote project path while preserving `~` expansion."""
-    if project_dir == "~":
-        return '"$HOME"'
-
-    if project_dir.startswith("~/"):
-        suffix = project_dir[2:].replace('"', '\\"')
-        return f'"$HOME/{suffix}"'
-
-    return shlex.quote(project_dir)
-
-
-def build_ssh_command(
-    config: RemoteConfig,
-    remote_command: str,
-    *,
-    tty: bool = False,
-) -> list[str]:
-    """Build one SSH command targeting the Raspberry Pi."""
-    wrapped_command = (
-        f"cd {quote_remote_project_dir(config.project_dir)} && {remote_command}"
-    )
-    ssh_command = ["ssh"]
-    if tty:
-        ssh_command.append("-t")
-    ssh_command.extend([config.ssh_target, f"bash -lc {shlex.quote(wrapped_command)}"])
-    return ssh_command
-
-
-def run_remote(config: RemoteConfig, remote_command: str, tty: bool = False) -> int:
-    """Execute one command on the Raspberry Pi via SSH."""
-    ssh_command = build_ssh_command(config, remote_command, tty=tty)
-
-    print("")
-    print(f"[pi-remote] host={config.ssh_target}")
-    print(f"[pi-remote] dir={config.project_dir}")
-    print(f"[pi-remote] cmd={remote_command}")
-    print("")
-
-    completed = subprocess.run(ssh_command, check=False)
-    return completed.returncode
-
-
-def run_remote_capture(
-    config: RemoteConfig,
-    remote_command: str,
-) -> subprocess.CompletedProcess[str]:
-    """Execute one SSH command and capture its stdout/stderr."""
-    ssh_command = build_ssh_command(config, remote_command)
-    return subprocess.run(
-        ssh_command,
-        check=False,
-        capture_output=True,
-        text=True,
-    )
-
-
-def run_local(command: Sequence[str], label: str) -> int:
-    """Execute one local command and stream its output."""
-    print("")
-    print(f"[pi-remote] local={label}")
-    print(f"[pi-remote] cmd={shlex.join(command)}")
-    print("")
-
-    completed = subprocess.run(list(command), check=False)
-    return completed.returncode
-
-
-def run_local_capture(command: Sequence[str]) -> subprocess.CompletedProcess[str]:
-    """Execute one local command and capture its stdout/stderr."""
-    return subprocess.run(
-        list(command),
-        check=False,
-        capture_output=True,
-        text=True,
-    )
-
-
-def shell_quote(value: str) -> str:
-    """Shell-escape one literal value for the remote command string."""
-
-    return shlex.quote(value)
-
-
-def _activate_script_path(venv: str) -> str:
-    """Return the shell path for activating the configured virtualenv."""
-    normalized = venv.rstrip("/")
-    if normalized.endswith("/bin/activate"):
-        return normalized
-    return f"{normalized}/bin/activate"
-
-
-def build_status_command(deploy_config: PiDeployConfig | None = None) -> str:
-    """Create the remote status command."""
-    deploy = deploy_config or load_pi_deploy_config()
-    return " && ".join(
-        [
-            "echo '== Git ==' ",
-            "git branch --show-current",
-            "git rev-parse --short HEAD",
-            "git status --short",
-            "echo",
-            "echo '== Music Backend ==' ",
-            "pgrep -af mpv || true",
-            "echo",
-            "echo '== YoyoPod Service ==' ",
-            "systemctl is-active \"yoyopod@$(id -un).service\" || true",
-            "echo",
-            "echo '== PiSugar Server ==' ",
-            "systemctl is-active pisugar-server || true",
-            "echo",
-            "echo '== PID File ==' ",
-            (
-                f"if test -f {shell_quote(deploy.pid_file)}; "
-                f"then cat {shell_quote(deploy.pid_file)}; "
-                "else echo 'missing'; fi"
-            ),
-            "echo",
-            "echo '== Latest Startup Marker ==' ",
-            (
-                f"if test -f {shell_quote(deploy.log_file)}; "
-                f"then grep -F {shell_quote(deploy.startup_marker)} {shell_quote(deploy.log_file)} | tail -n 1 || true; "
-                "else echo 'missing'; fi"
-            ),
-            "echo",
-            "echo '== Top Processes ==' ",
-            "ps -eo pid,comm,%mem,%cpu --sort=-%mem | head -15",
-        ]
-    )
-
-
-def build_sync_command(config: RemoteConfig, skip_uv_sync: bool) -> str:
-    """Create the remote sync command."""
-    commands = [
-        "git fetch origin",
-        f"git checkout {shlex.quote(config.branch)}",
-        f"git pull --ff-only origin {shlex.quote(config.branch)}",
-    ]
-    if not skip_uv_sync:
-        commands.append("uv sync --extra dev")
-    return " && ".join(commands)
-
-
-def build_rsync_command(
-    config: RemoteConfig,
-    deploy_config: PiDeployConfig,
-    *,
-    executable: str = "rsync",
-) -> list[str]:
-    """Create the local rsync command for a dirty-tree sync."""
-    command = [executable, "-avz", "--delete"]
-    for pattern in deploy_config.rsync_exclude:
-        command.extend(["--exclude", pattern])
-
-    remote_dir = config.project_dir.rstrip("/")
-    command.extend(["./", f"{config.ssh_target}:{remote_dir}/"])
-    return command
-
-
-def resolve_local_executable(program: str) -> str | None:
-    """Resolve one local executable, including common Windows install paths."""
-
-    resolved = shutil.which(program)
-    if resolved:
-        return resolved
-
-    if not sys.platform.startswith("win"):
-        return None
-
-    candidate_paths: list[Path] = []
-    if program.lower() == "rsync":
-        candidate_paths.extend(
-            [
-                Path(r"C:\msys64\usr\bin\rsync.exe"),
-                Path(r"C:\Program Files\Git\usr\bin\rsync.exe"),
-                Path(r"C:\Program Files\Git\bin\rsync.exe"),
-            ]
-        )
-    elif program.lower() == "scp":
-        candidate_paths.extend(
-            [
-                Path(r"C:\Windows\System32\OpenSSH\scp.exe"),
-            ]
-        )
-
-    for candidate in candidate_paths:
-        if candidate.exists():
-            return str(candidate)
-    return None
-
-
-def should_use_direct_rsync(rsync_binary: str | None) -> bool:
-    """Return whether the local rsync binary is safe for direct remote sync."""
-
-    if not rsync_binary:
-        return False
-
-    force_rsync = os.getenv("YOYOPOD_PI_FORCE_RSYNC", "").strip().lower()
-    if force_rsync in {"1", "true", "yes", "on"}:
-        return True
-
-    if not sys.platform.startswith("win"):
-        return True
-
-    normalized = str(Path(rsync_binary)).replace("/", "\\").lower()
-    known_windows_rsyncs = {
-        r"c:\msys64\usr\bin\rsync.exe",
-        r"c:\program files\git\usr\bin\rsync.exe",
-        r"c:\program files\git\bin\rsync.exe",
-    }
-    return normalized not in known_windows_rsyncs
-
-
-def sync_path_is_excluded(
-    rel_path: str,
-    patterns: Sequence[str],
-    *,
-    is_dir: bool,
-) -> bool:
-    """Return whether one repo-relative path should be skipped during sync."""
-
-    normalized = rel_path.strip("/")
-    if not normalized:
-        return False
-
-    segments = normalized.split("/")
-    basename = segments[-1]
-    for pattern in patterns:
-        candidate = str(pattern).strip()
-        if not candidate:
-            continue
-
-        if candidate.endswith("/"):
-            dir_pattern = candidate.rstrip("/")
-            if any(fnmatch.fnmatch(segment, dir_pattern) for segment in segments):
-                return True
-            continue
-
-        if fnmatch.fnmatch(normalized, candidate) or fnmatch.fnmatch(basename, candidate):
-            return True
-
-    return False
-
-
-def build_sync_file_manifest(
-    repo_root: Path,
-    deploy_config: PiDeployConfig,
-) -> list[str]:
-    """Collect the repo-relative file list that should be mirrored to the Pi."""
-
-    manifest: list[str] = []
-    for current_root, dirnames, filenames in os.walk(repo_root):
-        current_root_path = Path(current_root)
-        relative_root = current_root_path.relative_to(repo_root)
-
-        kept_dirnames: list[str] = []
-        for dirname in sorted(dirnames):
-            rel_dir = (relative_root / dirname).as_posix()
-            if rel_dir == ".":
-                rel_dir = dirname
-            if sync_path_is_excluded(rel_dir, deploy_config.rsync_exclude, is_dir=True):
-                continue
-            kept_dirnames.append(dirname)
-        dirnames[:] = kept_dirnames
-
-        for filename in sorted(filenames):
-            rel_file = (relative_root / filename).as_posix()
-            if rel_file == ".":
-                rel_file = filename
-            if sync_path_is_excluded(rel_file, deploy_config.rsync_exclude, is_dir=False):
-                continue
-            manifest.append(rel_file)
-
-    return manifest
-
-
-def build_archive_sync_extract_command(
-    config: RemoteConfig,
-    *,
-    archive_path: str,
-    manifest_path: str,
-) -> str:
-    """Create the remote command that unpacks and mirrors an scp-uploaded archive."""
-
-    project_dir_literal = repr(config.project_dir)
-    archive_path_literal = repr(archive_path)
-    manifest_path_literal = repr(manifest_path)
-    return f"""python - <<'PY'
-import fnmatch
-import json
-import os
-import tarfile
-from pathlib import Path
-
-
-def is_excluded(rel_path: str, patterns: tuple[str, ...]) -> bool:
-    normalized = rel_path.strip("/")
-    if not normalized:
-        return False
-
-    segments = normalized.split("/")
-    basename = segments[-1]
-    for pattern in patterns:
-        candidate = str(pattern).strip()
-        if not candidate:
-            continue
-
-        if candidate.endswith("/"):
-            dir_pattern = candidate.rstrip("/")
-            if any(fnmatch.fnmatch(segment, dir_pattern) for segment in segments):
-                return True
-            continue
-
-        if fnmatch.fnmatch(normalized, candidate) or fnmatch.fnmatch(basename, candidate):
-            return True
-
-    return False
-
-
-project_dir = Path(os.path.expanduser({project_dir_literal})).resolve()
-archive_path = Path({archive_path_literal})
-manifest_path = Path({manifest_path_literal})
-
-with manifest_path.open("r", encoding="utf-8") as handle:
-    payload = json.load(handle)
-
-expected_files = set(payload["files"])
-exclude_patterns = tuple(payload["exclude"])
-project_dir.mkdir(parents=True, exist_ok=True)
-
-with tarfile.open(archive_path, "r:gz") as archive:
-    members = archive.getmembers()
-    for member in members:
-        member_path = Path(member.name)
-        if member_path.is_absolute() or ".." in member_path.parts:
-            raise SystemExit(f"Unsafe archive member: {{member.name}}")
-    archive.extractall(project_dir)
-
-for path in sorted(project_dir.rglob("*"), key=lambda item: len(item.parts), reverse=True):
-    rel_path = path.relative_to(project_dir).as_posix()
-    if is_excluded(rel_path, exclude_patterns):
-        continue
-
-    if path.is_file() or path.is_symlink():
-        if rel_path not in expected_files:
-            path.unlink()
-        continue
-
-    try:
-        path.rmdir()
-    except OSError:
-        pass
-
-archive_path.unlink(missing_ok=True)
-manifest_path.unlink(missing_ok=True)
-PY"""
-
-
-def build_restart_command(deploy_config: PiDeployConfig) -> str:
-    """Create the remote restart command for the production app."""
-    pid_file = shell_quote(deploy_config.pid_file)
-    activate_script = shell_quote(_activate_script_path(deploy_config.venv))
-    service_name = 'yoyopod@"$(id -un)".service'
-
-    cleanup_commands = [
-        f"rm -f {pid_file}",
-    ]
-    for process_name in deploy_config.kill_processes:
-        cleanup_commands.append(
-            f"killall -9 {shell_quote(process_name)} >/dev/null 2>&1 || true"
-        )
-
-    cleanup_sequence = "; ".join(cleanup_commands)
-    manual_restart = (
-        f"(test -f {pid_file} && kill -9 $(cat {pid_file}) >/dev/null 2>&1) || true; "
-        "killall -9 python >/dev/null 2>&1 || true; "
-        f"{cleanup_sequence}; "
-        f"source {activate_script} && (nohup {deploy_config.start_cmd} > /dev/null 2>&1 &)"
-    )
-    managed_restart = (
-        f"if systemctl cat {service_name} >/dev/null 2>&1; then "
-        f"sudo systemctl stop {service_name} >/dev/null 2>&1 || true; "
-        f"{cleanup_sequence}; "
-        f"sudo systemctl start {service_name}; "
-        f"else {manual_restart}; "
-        "fi"
-    )
-    return " && ".join(
-        [
-            build_native_shim_refresh_command(deploy_config),
-            managed_restart,
-            build_startup_verification_command(deploy_config),
-        ]
-    )
-
-
-def build_native_shim_refresh_command(deploy_config: PiDeployConfig) -> str:
-    """Create the remote command that rebuilds missing or stale native shims."""
-
-    activate_script = shell_quote(_activate_script_path(deploy_config.venv))
-    return (
-        "{ "
-        f"source {activate_script} && python - <<'PY'\n"
-        "import subprocess\n"
-        "import sys\n"
-        "from pathlib import Path\n"
-        "\n"
-        "\n"
-        "def newest_mtime(paths: list[Path]) -> float:\n"
-        "    newest = 0.0\n"
-        "    for path in paths:\n"
-        "        if not path.exists():\n"
-        "            continue\n"
-        "        if path.is_dir():\n"
-        "            for candidate in path.rglob('*'):\n"
-        "                if candidate.is_file():\n"
-        "                    newest = max(newest, candidate.stat().st_mtime)\n"
-        "            continue\n"
-        "        newest = max(newest, path.stat().st_mtime)\n"
-        "    return newest\n"
-        "\n"
-        "\n"
-        "def is_stale(binary: Path, sources: list[Path]) -> bool:\n"
-        "    if not binary.exists():\n"
-        "        return True\n"
-        "    return newest_mtime(sources) > binary.stat().st_mtime\n"
-        "\n"
-        "\n"
-        "jobs = [\n"
-        "    (\n"
-        "        'LVGL',\n"
-        "        Path('yoyopy/ui/lvgl_binding/native/build/libyoyopy_lvgl_shim.so'),\n"
-        "        [Path('scripts/lvgl_build.py'), Path('yoyopy/ui/lvgl_binding/native')],\n"
-        "        [sys.executable, 'scripts/lvgl_build.py'],\n"
-        "    ),\n"
-        "    (\n"
-        "        'Liblinphone',\n"
-        "        Path('yoyopy/voip/liblinphone_binding/native/build/libyoyopy_liblinphone_shim.so'),\n"
-        "        [Path('scripts/liblinphone_build.py'), Path('yoyopy/voip/liblinphone_binding/native')],\n"
-        "        [sys.executable, 'scripts/liblinphone_build.py'],\n"
-        "    ),\n"
-        "]\n"
-        "\n"
-        "for label, output, sources, command in jobs:\n"
-        "    if not is_stale(output, sources):\n"
-        "        continue\n"
-        "    print(f'[pi-remote] info=rebuilding {label} native shim')\n"
-        "    subprocess.run(command, check=True)\n"
-        "PY\n"
-        "} "
-    )
-
-
-def build_smoke_command(args: argparse.Namespace) -> str:
-    """Create the remote smoke-validation command."""
-    parts = ["uv run python scripts/pi_smoke.py"]
-    if getattr(args, "with_power", False):
-        parts.append("--with-power")
-    if getattr(args, "with_rtc", False):
-        parts.append("--with-rtc")
-    if getattr(args, "with_music", False):
-        parts.append("--with-music")
-    if getattr(args, "with_voip", False):
-        parts.append("--with-voip")
-    if getattr(args, "with_lvgl_soak", False):
-        parts.append("--with-lvgl-soak")
-    if getattr(args, "verbose", False):
-        parts.append("--verbose")
-    if getattr(args, "music_timeout", 5) != 5:
-        parts.extend(["--music-timeout", str(args.music_timeout)])
-    if getattr(args, "voip_timeout", 10.0) != 10.0:
-        parts.extend(["--voip-timeout", str(args.voip_timeout)])
-    return " ".join(parts)
-
-
-def build_lvgl_soak_command(args: argparse.Namespace) -> str:
-    """Create the remote LVGL soak command."""
-
-    parts = ["uv run python scripts/lvgl_soak.py"]
-    if args.verbose:
-        parts.append("--verbose")
-    if args.cycles != 2:
-        parts.extend(["--cycles", str(args.cycles)])
-    if args.hold_seconds != 0.2:
-        parts.extend(["--hold-seconds", str(args.hold_seconds)])
-    if args.skip_sleep:
-        parts.append("--skip-sleep")
-    return " ".join(parts)
-
-
-def build_whisplay_command(args: argparse.Namespace) -> str:
-    """Create the remote Whisplay tuning command."""
-    parts = ["uv run python scripts/whisplay_tune.py"]
-    if args.verbose:
-        parts.append("--verbose")
-    if args.no_display:
-        parts.append("--no-display")
-    if args.duration_seconds != 30.0:
-        parts.extend(["--duration-seconds", str(args.duration_seconds)])
-    if args.debounce_ms is not None:
-        parts.extend(["--debounce-ms", str(args.debounce_ms)])
-    if args.double_tap_ms is not None:
-        parts.extend(["--double-tap-ms", str(args.double_tap_ms)])
-    if args.long_hold_ms is not None:
-        parts.extend(["--long-hold-ms", str(args.long_hold_ms)])
-    return " ".join(parts)
-
-
-def build_rtc_command(args: argparse.Namespace) -> str:
-    """Create the remote PiSugar RTC command."""
-    parts = ["uv run python scripts/pisugar_rtc.py"]
-    if args.verbose:
-        parts.append("--verbose")
-    parts.append(args.rtc_action)
-    if args.rtc_action == "set-alarm":
-        if not args.time:
-            raise SystemExit("--time is required for `pi_remote.py rtc set-alarm`")
-        parts.extend(["--time", shlex.quote(args.time)])
-        if args.repeat_mask != 127:
-            parts.extend(["--repeat-mask", str(args.repeat_mask)])
-    return " ".join(parts)
-
-
-def build_power_command(args: argparse.Namespace) -> str:
-    """Create the remote PiSugar power-status command."""
-    parts = ["uv run python scripts/pisugar_power.py"]
-    if args.verbose:
-        parts.append("--verbose")
-    return " ".join(parts)
-
-
-def build_logs_command(
-    args: argparse.Namespace,
-    deploy_config: PiDeployConfig | None = None,
-) -> str:
-    """Create the remote file-log inspection command."""
-
-    deploy = deploy_config or load_pi_deploy_config()
-    target_log = deploy.error_log_file if args.errors else deploy.log_file
-    tail_mode = "-F" if args.follow else ""
-    base_tail = f"tail -n {args.lines} {tail_mode} {shell_quote(target_log)}".strip()
-
-    if args.filter:
-        return (
-            f"test -f {shell_quote(target_log)} && "
-            f"{base_tail} | grep --line-buffered -i -- {shell_quote(args.filter)}"
-        )
-
-    return f"test -f {shell_quote(target_log)} && {base_tail}"
-
-
-def build_startup_verification_command(
-    deploy_config: PiDeployConfig | None = None,
-    *,
-    attempts: int = 20,
-) -> str:
-    """Create a remote command that waits for the startup marker and matching PID."""
-
-    deploy = deploy_config or load_pi_deploy_config()
-    pid_file = shell_quote(deploy.pid_file)
-    log_file = shell_quote(deploy.log_file)
-    startup_marker = shell_quote(deploy.startup_marker)
-    return " && ".join(
-        [
-            (
-                f"for _ in $(seq 1 {attempts}); do "
-                f"test -f {pid_file} && break; "
-                "sleep 1; "
-                "done"
-            ),
-            f"test -f {pid_file}",
-            f'pid="$(tr -d \'\\n\' < {pid_file})"',
-            'test -n "$pid"',
-            'kill -0 "$pid"',
-            (
-                f"for _ in $(seq 1 {attempts}); do "
-                f"if test -f {log_file} && "
-                f"grep -F {startup_marker} {log_file} | tail -n 1 | grep -F \"pid=$pid\" >/dev/null; then "
-                "break; "
-                "fi; "
-                "sleep 1; "
-                "done"
-            ),
-            f"grep -F {startup_marker} {log_file} | tail -n 1 | grep -F \"pid=$pid\"",
-        ]
-    )
-
-
-def build_service_command(
-    args: argparse.Namespace,
-    deploy_config: PiDeployConfig | None = None,
-) -> str:
-    """Create the remote systemd service command."""
-    deploy = deploy_config or load_pi_deploy_config()
-    service_name = 'yoyopod@"$(id -un)".service'
-    verify_startup = build_startup_verification_command(deploy)
-
-    if args.service_action == "status":
-        return f"sudo systemctl status {service_name} --no-pager || true"
-
-    if args.service_action == "install":
-        return " && ".join(
-            [
-                "test -f deploy/systemd/yoyopod@.service",
-                "sudo cp deploy/systemd/yoyopod@.service /etc/systemd/system/yoyopod@.service",
-                "sudo systemctl daemon-reload",
-                f"sudo systemctl enable --now {service_name}",
-                verify_startup,
-                f"sudo systemctl status {service_name} --no-pager",
-            ]
-        )
-
-    if args.service_action == "start":
-        return (
-            f"sudo systemctl start {service_name} && "
-            f"{verify_startup} && "
-            f"sudo systemctl status {service_name} --no-pager"
-        )
-
-    if args.service_action == "stop":
-        return f"sudo systemctl stop {service_name} && sudo systemctl status {service_name} --no-pager || true"
-
-    if args.service_action == "restart":
-        return (
-            f"sudo systemctl restart {service_name} && "
-            f"{verify_startup} && "
-            f"sudo systemctl status {service_name} --no-pager"
-        )
-
-    if args.service_action == "logs":
-        return f"sudo journalctl -u {service_name} -n {args.lines} --no-pager"
-
-    raise SystemExit(f"Unsupported service action: {args.service_action}")
-
-
-def build_run_command(args: argparse.Namespace) -> str:
-    """Create the remote production-app command."""
-    parts = ["uv run python yoyopod.py"]
-    if args.simulate:
-        parts.append("--simulate")
-    for app_arg in args.app_arg:
-        parts.append(shlex.quote(app_arg))
-    return " ".join(parts)
-
-
-def build_local_preflight_commands() -> list[tuple[str, list[str]]]:
-    """Create the local verification commands for preflight."""
-    return [
-        (
-            "compileall",
-            [
-                sys.executable,
-                "-m",
-                "compileall",
-                "yoyopy",
-                "tests",
-                "scripts/pi_smoke.py",
-                "scripts/pi_remote.py",
-                "scripts/pisugar_rtc.py",
-                "scripts/pisugar_power.py",
-                "scripts/whisplay_tune.py",
-                "scripts/lvgl_soak.py",
-            ],
-        ),
-        (
-            "pytest",
-            ["uv", "run", "pytest", "-q"],
-        ),
-    ]
-
-
-def run_rsync_deploy(
-    config: RemoteConfig,
-    deploy_config: PiDeployConfig,
-    *,
-    skip_restart: bool,
-) -> int:
-    """Rsync the local working tree to the Pi and optionally restart the app."""
-    rsync_binary = resolve_local_executable("rsync")
-    if should_use_direct_rsync(rsync_binary):
-        exit_code = run_local(
-            build_rsync_command(config, deploy_config, executable=rsync_binary),
-            "rsync",
-        )
-    else:
-        scp_binary = resolve_local_executable("scp")
-        if not scp_binary:
-            raise SystemExit(
-                "Neither `rsync` nor `scp` is available locally. "
-                "Install one of them or run from a machine with SSH copy tools."
-            )
-
-        print("")
-        if rsync_binary:
-            print(
-                "[pi-remote] info=local Windows rsync is not reliable for remote Unix paths; "
-                "falling back to archive+scp sync"
-            )
-        else:
-            print("[pi-remote] info=local rsync not found, falling back to archive+scp sync")
-        print("")
-
-        manifest = build_sync_file_manifest(REPO_ROOT, deploy_config)
-        remote_archive_path = "/tmp/yoyopod_sync.tar.gz"
-        remote_manifest_path = "/tmp/yoyopod_sync_manifest.json"
-
-        with tempfile.TemporaryDirectory(prefix="yoyopod-sync-") as temp_dir_raw:
-            temp_dir = Path(temp_dir_raw)
-            archive_path = temp_dir / "yoyopod_sync.tar.gz"
-            manifest_path = temp_dir / "yoyopod_sync_manifest.json"
-
-            with tarfile.open(archive_path, "w:gz") as archive:
-                for rel_path in manifest:
-                    archive.add(REPO_ROOT / rel_path, arcname=rel_path)
-
-            manifest_payload = {
-                "files": manifest,
-                "exclude": list(deploy_config.rsync_exclude),
-            }
-            manifest_path.write_text(
-                json.dumps(manifest_payload, indent=2, sort_keys=True),
-                encoding="utf-8",
-            )
-
-            upload_command = [
-                scp_binary,
-                str(archive_path),
-                str(manifest_path),
-                f"{config.ssh_target}:/tmp/",
-            ]
-            exit_code = run_local(upload_command, "scp-sync-upload")
-            if exit_code == 0:
-                exit_code = run_remote(
-                    config,
-                    build_archive_sync_extract_command(
-                        config,
-                        archive_path=remote_archive_path,
-                        manifest_path=remote_manifest_path,
-                    ),
-                )
-
-    if exit_code != 0 or skip_restart:
-        return exit_code
-
-    return run_remote(config, build_restart_command(deploy_config))
-
-
-def run_screenshot(
-    config: RemoteConfig,
-    deploy_config: PiDeployConfig,
-    args: argparse.Namespace,
-) -> int:
-    """Capture a screenshot from the remote app and copy it locally."""
-    pid_file = shell_quote(deploy_config.pid_file)
-    screenshot_path = shell_quote(deploy_config.screenshot_path)
-
-    alive_result = run_remote_capture(
-        config,
-        f"test -f {pid_file} && kill -0 $(cat {pid_file}) 2>/dev/null && echo ALIVE || echo DEAD",
-    )
-    if alive_result.returncode != 0 or alive_result.stdout.strip() != "ALIVE":
-        print("Remote app is not running; restart it before requesting a screenshot.")
-        if alive_result.stderr.strip():
-            print(alive_result.stderr.strip())
-        return 1
-
-    clear_result = run_remote_capture(
-        config,
-        f"rm -f {screenshot_path}",
-    )
-    if clear_result.returncode != 0:
-        print("Failed to clear the previous screenshot on the Raspberry Pi.")
-        if clear_result.stderr.strip():
-            print(clear_result.stderr.strip())
-        return clear_result.returncode
-
-    signal_name = "USR1" if args.readback else "USR2"
-    signal_result = run_remote_capture(
-        config,
-        f"kill -{signal_name} $(cat {pid_file})",
-    )
-    if signal_result.returncode != 0:
-        print("Failed to trigger screenshot capture on the Raspberry Pi.")
-        if signal_result.stderr.strip():
-            print(signal_result.stderr.strip())
-        return signal_result.returncode
-
-    verify_result = run_remote_capture(
-        config,
-        (
-            "for _ in $(seq 1 10); do "
-            f"test -f {screenshot_path} && echo READY && exit 0; "
-            "sleep 1; "
-            "done; "
-            "echo MISSING"
-        ),
-    )
-    if verify_result.returncode != 0 or verify_result.stdout.strip() != "READY":
-        print(
-            "Screenshot was not created on the Raspberry Pi. "
-            "Confirm the app is running and screenshot handlers are installed."
-        )
-        if verify_result.stderr.strip():
-            print(verify_result.stderr.strip())
-        return 1
-
-    output_path = Path(args.output).expanduser().resolve()
-    output_path.parent.mkdir(parents=True, exist_ok=True)
-    scp_command = [
-        "scp",
-        f"{config.ssh_target}:{deploy_config.screenshot_path}",
-        str(output_path),
-    ]
-    print("")
-    print(f"[pi-remote] local=screenshot-copy")
-    print(f"[pi-remote] cmd={shlex.join(scp_command)}")
-    print("")
-    copy_result = subprocess.run(scp_command, check=False)
-    if copy_result.returncode == 0:
-        print(f"Saved screenshot to {output_path}")
-    return copy_result.returncode
-
-
-def run_preflight(config: RemoteConfig, args: argparse.Namespace) -> int:
-    """Run the combined local + remote preflight flow."""
-    if not args.skip_local:
-        for label, command in build_local_preflight_commands():
-            exit_code = run_local(command, label)
-            if exit_code != 0:
-                return exit_code
-
-    if not args.skip_sync:
-        exit_code = run_remote(
-            config,
-            build_sync_command(config, args.skip_uv_sync),
-        )
-        if exit_code != 0:
-            return exit_code
-
-    return run_remote(config, build_smoke_command(args))
-
-
-def run_config_command(
-    args: argparse.Namespace,
-    deploy_config: PiDeployConfig,
-    *,
-    config_path: Path | None = None,
-    local_override_path: Path | None = None,
-) -> int:
-    """Show, create, or edit the local deploy config layers."""
-
-    base_path = config_path or DEPLOY_CONFIG_PATH
-    local_path = local_override_path or LOCAL_DEPLOY_CONFIG_PATH
-
-    if args.config_action == "show":
-        print(yaml.safe_dump(pi_deploy_config_to_dict(deploy_config), sort_keys=False).rstrip())
-        return 0
-
-    if args.config_action == "paths":
-        print(f"base: {base_path}")
-        print(f"local: {local_path}")
-        print(f"local_exists: {'yes' if local_path.exists() else 'no'}")
-        return 0
-
-    ensured_path, created = ensure_local_pi_deploy_config(
-        deploy_config,
-        local_override_path=local_path,
-    )
-
-    if args.config_action == "init-local":
-        state = "Created" if created else "Already exists"
-        print(f"{state}: {ensured_path}")
-        return 0
-
-    if args.config_action == "edit":
-        command = build_config_editor_command(
-            ensured_path,
-            editor=args.editor,
-        )
-        print("")
-        print(f"[pi-remote] local={'config-init' if created else 'config-edit'}")
-        print(f"[pi-remote] file={ensured_path}")
-        print(f"[pi-remote] cmd={shlex.join(command)}")
-        print("")
-        return subprocess.run(command, check=False).returncode
-
-    raise SystemExit(f"Unsupported config action: {args.config_action}")
-
-
-def main() -> int:
-    """Program entry point."""
-    deploy_config = load_pi_deploy_config()
-    parser = build_parser(deploy_config)
-    args = parser.parse_args()
-
-    if args.command == "config":
-        return run_config_command(args, deploy_config)
-
-    config = RemoteConfig(
-        host=args.host,
-        user=args.user,
-        project_dir=args.project_dir,
-        branch=args.branch,
-    )
-    validate_config(config)
-
-    if args.command == "status":
-        return run_remote(config, build_status_command(deploy_config))
-
-    if args.command == "sync":
-        return run_remote(config, build_sync_command(config, args.skip_uv_sync))
-
-    if args.command == "rsync":
-        return run_rsync_deploy(
-            config,
-            deploy_config,
-            skip_restart=args.skip_restart,
-        )
-
-    if args.command == "smoke":
-        return run_remote(config, build_smoke_command(args))
-
-    if args.command == "whisplay":
-        return run_remote(config, build_whisplay_command(args), tty=True)
-
-    if args.command == "lvgl-soak":
-        return run_remote(config, build_lvgl_soak_command(args), tty=True)
-
-    if args.command == "rtc":
-        return run_remote(config, build_rtc_command(args))
-
-    if args.command == "power":
-        return run_remote(config, build_power_command(args))
-
-    if args.command == "logs":
-        return run_remote(config, build_logs_command(args, deploy_config), tty=args.follow)
-
-    if args.command == "restart":
-        return run_remote(config, build_restart_command(deploy_config))
-
-    if args.command == "screenshot":
-        return run_screenshot(config, deploy_config, args)
-
-    if args.command == "service":
-        return run_remote(config, build_service_command(args, deploy_config))
-
-    if args.command == "preflight":
-        return run_preflight(config, args)
-
-    if args.command == "run":
-        return run_remote(config, build_run_command(args), tty=True)
-
-    parser.error(f"Unknown command: {args.command}")
-    return 2
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())
