+++ tests/test_whisplay_tune.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from scripts.whisplay_tune import apply_timing_overrides, summarize_timings
+from yoyopy.cli.pi.tune import apply_timing_overrides, summarize_timings
 
 
 def test_apply_timing_overrides_updates_input_section() -> None:
